+++ NeoForge/src/main/java/de/markusbordihn/modsoptimizer/ModsOptimizer.java
@@ -19,22 +19,10 @@
 
 package de.markusbordihn.modsoptimizer;
 
-import net.neoforged.fml.IExtensionPoint;
-import net.neoforged.fml.IExtensionPoint.DisplayTest;
-import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.common.Mod;
 
 @Mod(Constants.MOD_ID)
 public class ModsOptimizer {
 
-  public ModsOptimizer() {
-    // Make sure the mod being absent on the other network side does not cause the client to display
-    // the server as incompatible
-    ModLoadingContext.get()
-        .registerExtensionPoint(
-            IExtensionPoint.DisplayTest.class,
-            () ->
-                new IExtensionPoint.DisplayTest(
-                    () -> DisplayTest.IGNORESERVERONLY, (a, b) -> true));
-  }
+  public ModsOptimizer() {}
 }